+++ GarageManager.Services/Implementation/ServicesService.cs
@@ -1,5 +1,6 @@
 ﻿using GarageManager.Data.Context;
 using GarageManager.Data.Entities;
+using GarageManager.Services.Exceptions;
 using GarageManager.Services.Interfaces;
 using GarageManager.Services.SearchCriteria;
 using Microsoft.EntityFrameworkCore;
@@ -13,53 +14,83 @@ namespace GarageManager.Services.Implementation
 {
     public class ServicesService : IServicesService
     {
-        private readonly GarageManagerDbContext context;
+        private readonly GarageManagerDbContextFactory contextFactory;
 
-        public ServicesService(GarageManagerDbContext context)
+        public ServicesService(GarageManagerDbContextFactory contextFactory)
         {
-            this.context = context;
+            this.contextFactory = contextFactory;
         }
 
         public async Task CreateService(Service service)
         {
-            await context.Services.AddAsync(service);
-            await context.SaveChangesAsync();
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
+            {
+                await context.Services.AddAsync(service);
+                await context.SaveChangesAsync();
+            }
         }
 
         public async Task DeleteService(int serviceId)
         {
-            Service service = await context.Services.FindAsync(serviceId);
-            context.Services.Remove(service);
-            await context.SaveChangesAsync();
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
+            {
+                Service service = await context.Services.FindAsync(serviceId);
+
+                if (service == null)
+                {
+                    throw new ServiceNotFoundException(serviceId);
+                }
+
+                context.Services.Remove(service);
+                await context.SaveChangesAsync();
+            }
         }
 
         public async Task EditService(Service service)
         {
-            context.Services.Update(service);
-            await context.SaveChangesAsync();
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
+            {
+                context.Services.Update(service);
+                await context.SaveChangesAsync();
+            }
         }
 
         public async Task<Service> GetService(int serviceId)
         {
-            Service service = await context.Services.FindAsync(serviceId);
-            return service;
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
+            {
+                Service service = await context.Services.Include(s => s.Car).FirstOrDefaultAsync(s => s.ServiceId == serviceId);
+
+                if (service == null)
+                {
+                    throw new ServiceNotFoundException(serviceId);
+                }
+
+                return service;
+            }
         }
 
         public async Task<IEnumerable<Service>> GetServices()
         {
-            return await context.Services.ToListAsync();
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
+            {
+                return await context.Services.ToListAsync();
+            }
         }
 
         public async Task<IEnumerable<Service>> GetServices(ServicesListSearchCriteria servicesListSearchCriteria)
         {
-            IQueryable<Service> queryable = context.Services.AsQueryable();
-
-            if(!string.IsNullOrEmpty(servicesListSearchCriteria.Vin))
+            using (GarageManagerDbContext context = contextFactory.CreateDbContext())
             {
-                queryable = queryable.Where(s => s.Car.VIN.Equals(servicesListSearchCriteria.Vin));
-            }
+                IQueryable<Service> queryable = context.Services.AsQueryable();
 
-            return await queryable.ToListAsync();
+                if (!string.IsNullOrEmpty(servicesListSearchCriteria.Vin))
+                {
+                    queryable = queryable.Where(s => s.Car.Vin.Equals(servicesListSearchCriteria.Vin));
+                }
+
+                return await queryable.ToListAsync();
+            }
         }
     }
 }
